# Conflicts:
#	src/app/zap_cluster_list.json
#	src/controller/java/zap-generated/chip/devicecontroller/ChipIdLookup.java
#	src/darwin/Framework/CHIP/zap-generated/MTRAttributeSpecifiedCheck.mm
#	src/darwin/Framework/CHIP/zap-generated/MTRAttributeTLVValueDecoder.mm
#	src/darwin/Framework/CHIP/zap-generated/MTRBaseClusters.h
#	src/darwin/Framework/CHIP/zap-generated/MTRBaseClusters.mm
#	src/darwin/Framework/CHIP/zap-generated/MTRClusterConstants.h
#	src/darwin/Framework/CHIP/zap-generated/MTRClusters.h
#	src/darwin/Framework/CHIP/zap-generated/MTRClusters.mm
#	zzz_generated/darwin-framework-tool/zap-generated/cluster/Commands.h

+++ src/app/DefaultAttributePersistenceProvider.h
@@ -50,7 +50,7 @@ public:
 
     // AttributePersistenceProvider implementation.
     CHIP_ERROR WriteValue(const ConcreteAttributePath & aPath, const ByteSpan & aValue) override;
-    CHIP_ERROR ReadValue(const ConcreteAttributePath & aPath, EmberAfAttributeType aType, uint16_t aSize,
+    CHIP_ERROR ReadValue(const ConcreteAttributePath & aPath, EmberAfAttributeType aType, size_t aSize,
                          MutableByteSpan & aValue) override;
 
 protected:
